* commit 'f5b553d0eb9327b1bcaea7b70f35cdf5c9466acb': (100 commits)
  Initial support for photogrammetry
  I forgot to filter unicode characters on query input.
  Ignore special characters on object search
  Added option to hide a specific category
  Images ordering on object view bugfix
  Celery config minor bug fixes
  Yet another try to speed up vision api search
  Minor bug fixespt3
  Minor bug fixes pt2
  Minor bug fixes pt2
  Minor bug fixes
  Celery priorities changed
  Visionapi speed optimization
  Minor bug fixes
  Vision tests bugfix
  Optimized vision_api results
  Like count automatic update on toggle_like
  Things files filter integration
  Things filter integration
  I forgot to add the property decorator in the last commit :)
  ...

+++ db/tools/stl_to_sfb.py
@@ -9,7 +9,7 @@ from django.core.files.base import ContentFile
 import random, string
 import gc
 
-def convert(stlpath,name):
+def convert(stlpath,name,rotate=False):
     '''
     El proceso de conversion a sfb consiste en
     1) Reescalar el stl con un factor 1/1000, ya que sfb asume que las unidades de los modelos están en metros
@@ -26,15 +26,29 @@ def convert(stlpath,name):
         raise Exception("Error al importar STL a trimesh")
     t = trimesh.transformations.scale_matrix(1/1000, [0,0,0])
     mesh.apply_transform(t)
-    #Guardamos el OBJ
+    if rotate:
+        t = trimesh.transformations.rotation_matrix(-np.pi/2, [1,0,0])
+        mesh.apply_transform(t)
+    #Guardamos el STL transformado
+    stl_path = settings.BASE_DIR + '/tmp/' + name.split('/')[-1].split('.')[0] + '.stl'
+    with open(stl_path,'wb') as f:
+        f.write(trimesh.io.stl.export_stl(mesh))
+    #Convertimos el stl a obj usando assimp
     obj_path = settings.BASE_DIR + '/tmp/' + name.split('/')[-1].split('.')[0] + '.obj'
-    with open(obj_path,'w') as f:
-        f.write(trimesh.io.wavefront.export_wavefront(mesh))
+    args = ['assimp', 'export', stl_path, obj_path]
+    proc = subprocess.run(args,universal_newlines = True, stdout = subprocess.PIPE, stderr = subprocess.PIPE)
+    #Se convirtio exitosamente?
+    if 'Exporting file ...                   OK ' not in proc.stdout.splitlines():
+        print(proc.stdout.splitlines())
+        raise Exception("Error al convertir STL a OBJ")
     #Convertimos el sfb
     args = [settings.BASE_DIR + '/lib/sceneform_sdk/linux/converter','-d','--mat',settings.BASE_DIR + '/lib/sceneform_sdk/default_materials/obj_material.sfm',
     '--outdir', settings.BASE_DIR + '/tmp/',obj_path]
     proc = subprocess.run(args,universal_newlines = True, stdout = subprocess.PIPE, stderr = subprocess.PIPE)
+    #Borramos los archivos generados
     os.remove(obj_path)
+    os.remove(stl_path)
+    os.remove(obj_path+'.mtl')
     #Se proceso correctamente?
     for line in proc.stdout.splitlines():
         if 'Wrote SFB to' in line:
@@ -75,6 +89,13 @@ def combine_stls_files(objeto):
             args_d = args.copy() + ['-x',str(build_plate_size),'-y',str(build_plate_size),'--spacing',str(10),'--outputdir',dir+'/']
             os.mkdir(dir)
             proc = subprocess.run(args_d,universal_newlines = True, stdout = subprocess.PIPE, stderr = subprocess.PIPE,cwd=dir+'/')
+            for line in proc.stdout.splitlines():
+                if 'Generating plate 0' in line:
+                    break
+            else:
+                print("Error al acomodar las piezas para el modelo sfb")
+                print(proc.stdout.splitlines())
+                print(proc.stderr.splitlines())
             return dir
 
 def combine_stl_with_correct_coordinates(objeto):
